// appAngular/src/app/models/formulario.ts
export class Formulario {
  id: any;
  nome: string;
  primeiroCadastro: string;
  rg: string;
  cpf: any;
  nomeMae: string;
  nomePai: string;
  dataNascimento: any;
  sexo: string;
  cep: any;
  numero: any;
  complemento: string;
  uf: string;
  cidade: string;
  endereco: string;
  bairro: string;
  telefone: string;
  celular: string;
  email: string;
  escolaridade: string;
  estadoCivil: string;
  profissao: string;
  foto: any;
  anexos: any;
  atendenteCpf: any;
}